ISSUE #2865 user agent helper methods moved in sessions service respond method no longed accepts username as param and logout passes the username through req
