+++ backend/src/v5/utils/helper/strings.js
@@ -16,9 +16,6 @@
  */
 
 const { camelCase, snakeCase } = require('lodash');
-const Device = require('device');
-const UaParserJs = require('ua-parser-js');
-const geoip = require('geoip-lite');
 
 const StringHelper = {};
 // Turns thisIsUs to THIS_IS_US
@@ -37,52 +34,5 @@ StringHelper.hasEmailFormat = (str) => {
 	const emailRegex = /^(?:[a-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\.)+[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-zA-Z0-9-]*[a-zA-Z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])$/;
 	return emailRegex.test(String(str).toLowerCase());
 };
-
-// Format:
-// PLUGIN: {OS Name}/{OS Version} {Host Software Name}/{Host Software Version} {Plugin Type}/{Plugin Version}
-// Example:
-// PLUGIN: Windows/10.0.19042.0 REVIT/2021.1 PUBLISH/4.15.0
-StringHelper.getUserAgentInfoFromPlugin = (userAgentString) => {
-	const [osInfo, appInfo, engineInfo] = userAgentString.replace('PLUGIN: ', '').split(' ');
-
-	const osInfoComponents = osInfo.split('/');
-	const appInfoComponents = appInfo.split('/');
-	const engineInfoComponents = engineInfo.split('/');
-
-	const userAgentInfo = {
-		application: {
-			name: appInfoComponents[0],
-			version: appInfoComponents[1],
-			type: 'plugin',
-		},
-		engine: {
-			name: '3drepoplugin',
-			version: engineInfoComponents[1],
-		},
-		os: {
-			name: osInfoComponents[0],
-			version: osInfoComponents[1],
-		},
-		device: 'desktop',
-	};
-
-	return userAgentInfo;
-};
-
-StringHelper.getUserAgentInfoFromBrowser = (userAgentString) => {
-	const { browser, engine, os } = UaParserJs(userAgentString);
-	const userAgentInfo = {
-		application: browser.name ? { ...browser, type: 'browser' } : { type: 'unknown' },
-		engine,
-		os,
-		device: userAgentString ? Device(userAgentString).type : 'unknown',
-	};
-
-	return userAgentInfo;
-};
-
-StringHelper.isUserAgentFromPlugin = (userAgent) => userAgent.split(' ')[0] === 'PLUGIN:';
-
-StringHelper.getLocationFromIPAddress = (ipAddress) => geoip.lookup(ipAddress);
 
 module.exports = StringHelper;
